# artifacts/api-server/src/routes/estimates.ts
import { Router, type IRouter } from "express";
import { and, desc, eq } from "drizzle-orm";
import { db, estimatesTable, type Estimate } from "@workspace/db";
import {
  PatchEstimateBody,
  PatchEstimateParams,
  CreateEstimateBody,
  CreateEstimateResponse,
  GetEstimateParams,
  GetEstimateResponse,
  GetEstimateStatsResponse,
  ListAssetTypesResponse,
  ListEstimatesResponse,
  ListRegionsResponse,
} from "@workspace/api-zod";
import type { AssetType, EstimateInput, EstimateReport, EstimateResult } from "@workspace/api-zod";
import { ASSET_TYPES, getAssetType } from "../lib/assetTypes";
import { REGIONS, getRegion } from "../lib/regions";
import { generateEstimate } from "../lib/estimate";
import { sanitizeComparables } from "../lib/comparables";
import { requireAuth, getUserId, type AuthedRequest } from "../middlewares/requireAuth";
import { recordPlatformEvent } from "../lib/platformEvents";
import { getFxRateSnapshot } from "../lib/fxRates";
import { portfolioShelfFromEstimate, readAttributesFromStoredResult } from "@workspace/asset-shelf-tier";
import { logger } from "../lib/logger";
import { notifyEstimateReadyEmail } from "../lib/estimateReadyEmail";
import {
  FREE_MONTHLY_VALUATION_CAP,
  incrementMonthlyEstimateUsage,
  resolveUserEntitlements,
  valuationTierForEstimate,
  includeSellerPlaybookInEstimate,
} from "../lib/entitlements";
import { getPortfolioByIdForUser, resolveDefaultPortfolioId } from "../lib/portfoliosService";
import { computeEstimateStatsFromRows } from "../lib/estimateStatsRollup";

const router: IRouter = Router();

function isCompleteAssetType(v: unknown): v is AssetType {
  if (!v || typeof v !== "object") return false;
  const o = v as Record<string, unknown>;
  return (
    typeof o.id === "string" &&
    typeof o.name === "string" &&
    typeof o.category === "string" &&
    typeof o.tagline === "string" &&
    Array.isArray(o.fields) &&
    typeof o.exampleAttributes === "string" &&
    typeof o.internationallyTradeable === "boolean"
  );
}

function resolveEstimateAssetType(
  row: { assetTypeId: string; assetTypeName: string },
  embedded: unknown,
): AssetType {
  if (isCompleteAssetType(embedded)) return embedded;
  const fromRegistry = getAssetType(row.assetTypeId);
  if (fromRegistry) return fromRegistry;
  return {
    id: row.assetTypeId,
    name: row.assetTypeName,
    category: "Other",
    tagline: "",
    fields: [],
    exampleAttributes: "",
    internationallyTradeable: true,
  };
}

function isCompleteEstimateInput(v: unknown): v is EstimateInput {
  if (!v || typeof v !== "object") return false;
  const o = v as Record<string, unknown>;
  return (
    typeof o.assetTypeId === "string" &&
    typeof o.title === "string" &&
    typeof o.currentRegion === "string" &&
    typeof o.currency === "string" &&
    typeof o.condition === "number"
  );
}

function resolveEstimateInput(
  row: {
    assetTypeId: string;
    title: string;
    currency: string;
    bestArbitrageRegion: string;
  },
  embedded: unknown,
): EstimateInput {
  if (isCompleteEstimateInput(embedded)) return embedded;
  const partial =
    embedded && typeof embedded === "object" ? (embedded as Partial<EstimateInput>) : {};
  return {
    assetTypeId: partial.assetTypeId ?? row.assetTypeId,
    title: partial.title ?? row.title,
    currentRegion: partial.currentRegion ?? row.bestArbitrageRegion,
    currency: partial.currency ?? row.currency,
    condition: typeof partial.condition === "number" ? partial.condition : 5,
    brand: partial.brand,
    model: partial.model,
    year: partial.year,
    purchasePrice: partial.purchasePrice,
    attributes: partial.attributes,
    extraFields: partial.extraFields,
  };
}

function finiteNum(v: unknown, fallback: number): number {
  return typeof v === "number" && Number.isFinite(v) ? v : fallback;
}

function safeArray<T>(v: unknown): T[] {
  return Array.isArray(v) ? (v as T[]) : [];
}

function isCompleteEstimateReport(v: unknown): v is EstimateReport {
  if (!v || typeof v !== "object") return false;
  const o = v as Record<string, unknown>;
  return (
    typeof o.headline === "string" &&
    typeof o.summary === "string" &&
    typeof o.baselineNarrative === "string" &&
    typeof o.marketNarrative === "string" &&
    typeof o.arbitrageNarrative === "string" &&
    typeof o.worldEventsNarrative === "string" &&
    typeof o.finalNarrative === "string"
  );
}

function resolveEstimateReport(embedded: unknown, title: string): EstimateReport {
  if (isCompleteEstimateReport(embedded)) return embedded;
  const partial =
    embedded && typeof embedded === "object" ? (embedded as Partial<EstimateReport>) : {};
  return {
    headline: partial.headline ?? `Valuation · ${title}`,
    summary:
      partial.summary ??
      "This valuation was recovered from older data; some narrative sections may be missing.",
    baselineNarrative: partial.baselineNarrative ?? "",
    marketNarrative: partial.marketNarrative ?? "",
    arbitrageNarrative: partial.arbitrageNarrative ?? "",
    worldEventsNarrative: partial.worldEventsNarrative ?? "",
    finalNarrative: partial.finalNarrative ?? "",
  };
}

function mergeEstimateResultFromRow(row: Estimate, storedUnknown: unknown): EstimateResult {
  const raw =
    storedUnknown && typeof storedUnknown === "object"
      ? (storedUnknown as Record<string, unknown>)
      : {};
  const baselineMid = finiteNum(raw.baselineMid, row.baselineMid);
  const adjustedMid = finiteNum(raw.adjustedMid, row.adjustedMid);
  const tier: "free" | "pro" =
    raw.tier === "pro" || raw.tier === "free" ? raw.tier : (row.tier as "free" | "pro");

  return {
    ...raw,
    input: resolveEstimateInput(row, raw.input),
    assetType: resolveEstimateAssetType(row, raw.assetType),
    report: resolveEstimateReport(raw.report, row.title),
    marketSignals: safeArray(raw.marketSignals),
    worldEvents: safeArray(raw.worldEvents),
    arbitrage: safeArray(raw.arbitrage),
    comparables: sanitizeComparables(safeArray(raw.comparables)),
    netMarketFactor: finiteNum(raw.netMarketFactor, 1),
    currency: typeof raw.currency === "string" && raw.currency ? raw.currency : row.currency,
    tier,
    baselineLow: finiteNum(raw.baselineLow, baselineMid),
    baselineMid,
    baselineHigh: finiteNum(raw.baselineHigh, baselineMid),
    adjustedLow: finiteNum(raw.adjustedLow, adjustedMid),
    adjustedMid,
    adjustedHigh: finiteNum(raw.adjustedHigh, adjustedMid),
    bestArbitrageRegion:
      typeof raw.bestArbitrageRegion === "string" && raw.bestArbitrageRegion
        ? raw.bestArbitrageRegion
        : row.bestArbitrageRegion,
    id: row.id,
    createdAt: row.createdAt.toISOString(),
    intent:
      row.intent === "hold" || row.intent === "monitor" || row.intent === "sell"
        ? (row.intent as NonNullable<EstimateResult["intent"]>)
        : null,
  } as unknown as EstimateResult;
}

function readSellerRegionFromStoredResult(storedUnknown: unknown, fallback: string): string {
  const raw =
    storedUnknown && typeof storedUnknown === "object" ? (storedUnknown as Record<string, unknown>) : {};
  const inp =
    raw.input && typeof raw.input === "object" ? (raw.input as Record<string, unknown>) : undefined;
  const cr = inp?.currentRegion;
  return typeof cr === "string" && cr.trim().length > 0 ? cr.trim() : fallback;
}

router.get("/asset-types", async (_req, res): Promise<void> => {
  res.json(ListAssetTypesResponse.parse(ASSET_TYPES));
});

router.get("/regions", async (_req, res): Promise<void> => {
  res.json(ListRegionsResponse.parse(REGIONS));
});

router.get("/estimates", async (req, res): Promise<void> => {
  const userId = getUserId(req);
  if (!userId) {
    res.json(ListEstimatesResponse.parse([]));
    return;
  }
  const rows = await db
    .select()
    .from(estimatesTable)
    .where(eq(estimatesTable.userId, userId))
    .orderBy(desc(estimatesTable.createdAt))
    .limit(50);
  const summaries = rows.map((r) => {
    const attrs = readAttributesFromStoredResult(r.result);
    const portfolioShelf = portfolioShelfFromEstimate(attrs, r.assetTypeId);
    return {
      id: r.id,
      title: r.title,
      assetTypeId: r.assetTypeId,
      assetTypeName: r.assetTypeName,
      baselineMid: r.baselineMid,
      adjustedMid: r.adjustedMid,
      currency: r.currency,
      bestArbitrageRegion: r.bestArbitrageRegion,
      tier: r.tier === "pro" ? "pro" : "free",
      currentRegion: readSellerRegionFromStoredResult(r.result, r.bestArbitrageRegion),
      portfolioShelf,
      createdAt: r.createdAt.toISOString(),
      portfolioId: r.portfolioId ?? null,
      intent: r.intent === "hold" || r.intent === "monitor" || r.intent === "sell" ? r.intent : null,
    };
  });
  res.json(ListEstimatesResponse.parse(summaries));
});

router.get("/estimates/stats", async (req, res): Promise<void> => {
  const userId = getUserId(req);
  const rows = userId
    ? await db.select().from(estimatesTable).where(eq(estimatesTable.userId, userId))
    : [];
  const fx = await getFxRateSnapshot();
  const mult = fx.rates;
  const statsPayload = computeEstimateStatsFromRows(
    rows.map((r) => ({
      baselineMid: r.baselineMid,
      adjustedMid: r.adjustedMid,
      currency: r.currency,
      assetTypeName: r.assetTypeName,
      bestArbitrageRegion: r.bestArbitrageRegion,
    })),
    mult,
  );

  res.json(GetEstimateStatsResponse.parse(statsPayload));
});

router.post("/estimates", requireAuth, async (req, res): Promise<void> => {
  const userId = (req as AuthedRequest).userId!;
  const ent = await resolveUserEntitlements(userId);

  const body = CreateEstimateBody.safeParse(req.body);
  if (!body.success) {
    res.status(400).json({ error: body.error.message });
    return;
  }

  const limit = ent.valuationsMonthLimit ?? null;
  if (limit != null && ent.valuationsThisMonth >= limit) {
    res.status(429).json({
      error: `You've used your ${FREE_MONTHLY_VALUATION_CAP} free valuations this month. Upgrade an Everyday subscription for unlimited personal valuations.`,
    });
    return;
  }

  const assetType = getAssetType(body.data.assetTypeId);
  if (!assetType) {
    res.status(400).json({ error: "Unknown assetTypeId" });
    return;
  }

  let portfolioIdToUse: string;
  const requestedPf = body.data.portfolioId;
  if (typeof requestedPf === "string" && requestedPf.trim() !== "") {
    const pfRow = await getPortfolioByIdForUser(requestedPf, userId);
    if (!pfRow) {
      res.status(400).json({ error: "portfolioId must reference one of your portfolios." });
      return;
    }
    portfolioIdToUse = pfRow.id;
  } else {
    portfolioIdToUse = await resolveDefaultPortfolioId(userId);
  }

  const region = getRegion(body.data.currentRegion);
  const tier = valuationTierForEstimate(ent);
  const includePlaybook = includeSellerPlaybookInEstimate(ent);

  const input: EstimateInput = {
    ...(body.data as EstimateInput),
    assetTypeId: assetType.id,
    currency: region?.currencyCode ?? body.data.currency ?? "USD",
    portfolioId: portfolioIdToUse,
  };

  const computed = await generateEstimate(input, assetType, tier, includePlaybook);

  const [row] = await db
    .insert(estimatesTable)
    .values({
      userId,
      portfolioId: portfolioIdToUse,
      assetTypeId: assetType.id,
      assetTypeName: assetType.name,
      title: input.title,
      currency: computed.currency,
      baselineMid: computed.baselineMid,
      adjustedMid: computed.adjustedMid,
      bestArbitrageRegion: computed.bestArbitrageRegion,
      tier,
      result: computed,
    })
    .returning();

  await incrementMonthlyEstimateUsage(userId);

  const result = {
    ...computed,
    id: row.id,
    createdAt: row.createdAt.toISOString(),
  } as unknown as EstimateResult;
  void recordPlatformEvent({
    userId,
    eventType: "estimate.created",
    payload: {
      estimateId: row.id,
      assetTypeId: assetType.id,
      tier,
      adjustedMid: computed.adjustedMid,
      baselineMid: computed.baselineMid,
      currency: computed.currency,
      region: input.currentRegion,
    },
  });
  void notifyEstimateReadyEmail(userId, {
    id: row.id,
    title: input.title,
    assetTypeName: assetType.name,
  }).catch((err) => {
    logger.error({ err }, "notifyEstimateReadyEmail failed");
  });
  res.json(CreateEstimateResponse.parse(result));
});

router.patch("/estimates/:id", requireAuth, async (req, res): Promise<void> => {
  const params = PatchEstimateParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const parsedBody = PatchEstimateBody.safeParse(req.body);
  if (!parsedBody.success) {
    res.status(400).json({ error: parsedBody.error.message });
    return;
  }
  const userId = (req as AuthedRequest).userId!;
  const [updated] = await db
    .update(estimatesTable)
    .set({ intent: parsedBody.data.intent })
    .where(and(eq(estimatesTable.id, params.data.id), eq(estimatesTable.userId, userId)))
    .returning();
  if (!updated) {
    res.status(404).json({ error: "Estimate not found" });
    return;
  }
  const merged = mergeEstimateResultFromRow(updated, updated.result);
  res.json(GetEstimateResponse.parse(merged));
});

router.get("/estimates/:id", requireAuth, async (req, res): Promise<void> => {
  const params = GetEstimateParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const userId = (req as AuthedRequest).userId!;
  const [row] = await db
    .select()
    .from(estimatesTable)
    .where(and(eq(estimatesTable.id, params.data.id), eq(estimatesTable.userId, userId)));
  if (!row) {
    res.status(404).json({ error: "Estimate not found" });
    return;
  }
  const merged = mergeEstimateResultFromRow(row, row.result);
  res.json(GetEstimateResponse.parse(merged));
});

export default router;