# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: "ValYoued: multi-asset valuation API"
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /estimates:
    get:
      operationId: listEstimates
      tags: [estimate]
      summary: List recent estimates
      responses:
        "200":
          description: Recent estimates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EstimateSummary"
    post:
      operationId: createEstimate
      tags: [estimate]
      summary: Create a new asset estimate (tier derived from subscription / free limits)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateInput"
      responses:
        "200":
          description: Estimate result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateResult"
        "429":
          description: Free tier monthly valuation limit reached
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: string
  /estimates/{id}:
    get:
      operationId: getEstimate
      tags: [estimate]
      summary: Get an estimate by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateResult"
    patch:
      operationId: patchEstimate
      tags: [estimate]
      summary: Update estimate metadata (intent)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchEstimateBody"
      responses:
        "200":
          description: Updated estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateResult"
...
components:
  schemas:
...
    AssetField:
      type: object
      required: [key, label, type, required]
      properties:
        key:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text, number, select, textarea]
        required:
          type: boolean
        placeholder:
          type: string
        help:
          type: string
        options:
          type: array
          items:
            type: string
    AssetType:
      type: object
      required: [id, name, category, tagline, fields, exampleAttributes, internationallyTradeable]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          description: Group label (such as "Watches & Jewelry", "Vehicles", "Real Estate")
        tagline:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/AssetField"
        exampleAttributes:
          type: string
        internationallyTradeable:
          type: boolean
          description: Whether this asset is practical to ship/sell across borders
    EstimateInput:
      type: object
      required: [assetTypeId, title, condition, currentRegion, currency]
      properties:
        assetTypeId:
          type: string
        title:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        condition:
          type: number
        purchasePrice:
          type: number
          description: Original purchase price in the seller's local currency
        currentRegion:
          type: string
        currency:
          type: string
          description: ISO 4217 currency code derived from currentRegion
        attributes:
          type: string
        extraFields:
          type: object
          additionalProperties:
            type: string
          description: Free-form key/value attributes captured from the dynamic form
        portfolioId:
          type: string
          format: uuid
          description: Workspace to attach this valuation to; defaults to primary portfolio
    PatchEstimateBody:
      type: object
      additionalProperties: false
      properties:
        intent:
          type: string
          enum: [hold, monitor, sell]
      required: [intent]
...
    Comparable:
      type: object
      required: [source, description, price, year]
      properties:
        source:
          type: string
        description:
          type: string
        price:
          type: number
          description: Price in the result currency
        year:
          type: integer
        url:
          type: string
        conditionCue:
          type: string
          description: Condition / completeness shorthand for this comp (not the seller copy)
        locationOrChannel:
          type: string
          description: Geography or outlet shorthand (EU consignment, US auction floor, Japan Yahoo, etc.)
        transactionTypeGuess:
          type: string
          enum: [sold_estimate, asking_price, unknown]
        relevanceExplanation:
          type: string
          description: One concise line tying this comp to the wizard inputs or tier hint
        matchTier:
          type: string
          enum: [strong, moderate, broadAnalogue]
        imageUrl:
          type: string
          description: Optional HTTPS thumbnail when the cited evidence page has a stable image URL
    MarketSignal:
      type: object
      required: [label, value, impact, rationale]
      properties:
        label:
          type: string
        value:
          type: string
        impact:
          type: number
        rationale:
          type: string
    WorldEvent:
      type: object
      required: [title, summary, sentiment, scope]
      properties:
        title:
          type: string
        summary:
          type: string
        sentiment:
          type: string
          enum: [positive, negative, neutral]
        scope:
          type: string
          description: 'Typical values: "Global", "United Kingdom", "EU"'
        source:
          type: string
        url:
          type: string
        publishedAt:
          type: string
          description: Article publication date (RFC 2822 or ISO 8601)
    ArbitrageOption:
      type: object
      required: [region, marketplace, estimatedSalePrice, estimatedShipping, estimatedFees, estimatedDuties, netToSeller, currency, demandNote, recommended]
      properties:
        region:
          type: string
        marketplace:
          type: string
        estimatedSalePrice:
          type: number
        estimatedShipping:
          type: number
        estimatedFees:
          type: number
        estimatedDuties:
          type: number
        netToSeller:
          type: number
        currency:
          type: string
          description: Currency for the values in this row (matches result currency)
        demandNote:
          type: string
        recommended:
          type: boolean
    NegotiationTactic:
      type: object
      required: [title, detail]
      properties:
        title:
          type: string
        detail:
          type: string
    ProInsights:
      type: object
      required: [negotiationTactics, talkingPoints, redFlags, optimalTiming, listingTips, walkAwayPrice, anchorPrice]
      properties:
        negotiationTactics:
          type: array
          items:
            $ref: "#/components/schemas/NegotiationTactic"
        talkingPoints:
          type: array
          items:
            type: string
        redFlags:
          type: array
          items:
            type: string
        optimalTiming:
          type: string
        listingTips:
          type: array
          items:
            type: string
        walkAwayPrice:
          type: number
        anchorPrice:
          type: number
    EstimateReport:
      type: object
      required: [headline, summary, baselineNarrative, marketNarrative, arbitrageNarrative, worldEventsNarrative, finalNarrative]
      properties:
        headline:
          type: string
        summary:
          type: string
        baselineNarrative:
          type: string
        marketNarrative:
          type: string
        arbitrageNarrative:
          type: string
        worldEventsNarrative:
          type: string
        finalNarrative:
          type: string
    EstimateResult:
      type: object
      required:
        - id
        - createdAt
        - input
        - assetType
        - currency
        - baselineLow
        - baselineHigh
        - baselineMid
        - comparables
        - marketSignals
        - worldEvents
        - netMarketFactor
        - adjustedLow
        - adjustedHigh
        - adjustedMid
        - arbitrage
        - bestArbitrageRegion
        - report
        - tier
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        input:
          $ref: "#/components/schemas/EstimateInput"
        assetType:
          $ref: "#/components/schemas/AssetType"
        currency:
          type: string
        baselineLow:
          type: number
        baselineHigh:
          type: number
        baselineMid:
          type: number
        comparables:
          type: array
          items:
            $ref: "#/components/schemas/Comparable"
        marketSignals:
          type: array
          items:
            $ref: "#/components/schemas/MarketSignal"
        worldEvents:
          type: array
          items:
            $ref: "#/components/schemas/WorldEvent"
        netMarketFactor:
          type: number
        adjustedLow:
          type: number
        adjustedHigh:
          type: number
        adjustedMid:
          type: number
        arbitrage:
          type: array
          items:
            $ref: "#/components/schemas/ArbitrageOption"
        bestArbitrageRegion:
          type: string
        report:
          $ref: "#/components/schemas/EstimateReport"
        tier:
          type: string
          enum: [free, pro]
        proInsights:
          $ref: "#/components/schemas/ProInsights"
        intent:
          type: string
          nullable: true
          enum: [hold, monitor, sell]
    EstimateSummary:
      type: object
      required:
        [
          id,
          title,
          assetTypeId,
          assetTypeName,
          baselineMid,
          adjustedMid,
          currency,
          bestArbitrageRegion,
          tier,
          portfolioShelf,
          createdAt,
        ]
      properties:
        id:
          type: string
        title:
          type: string
        assetTypeId:
          type: string
        assetTypeName:
          type: string
        baselineMid:
          type: number
        adjustedMid:
          type: number
        currency:
          type: string
        bestArbitrageRegion:
          type: string
        tier:
          type: string
          enum: [free, pro]
          description: Tier used when the valuation was generated (Pro unlocks full arbitrage rows on the report).
        currentRegion:
          type: string
          description: Seller region from stored valuation input when available.
        portfolioShelf:
          type: string
          enum: [luxury, everyday, other]
          description: >-
            Shelf for portfolio grouping: from the seller tier captured on the estimate plus asset-class hints.
        createdAt:
          type: string
          format: date-time
        portfolioId:
          type: string
          format: uuid
          nullable: true
        intent:
          type: string
          nullable: true
          enum: [hold, monitor, sell]
...