# lib/api-zod/src/generated/types/comparable.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ValYoued: multi-asset valuation API
 * OpenAPI spec version: 0.1.0
 */
import type { ComparableMatchTier } from "./comparableMatchTier";
import type { ComparableTransactionTypeGuess } from "./comparableTransactionTypeGuess";

export interface Comparable {
  source: string;
  description: string;
  /** Price in the result currency */
  price: number;
  year: number;
  url?: string;
  /** Condition / completeness shorthand for this comp (not the seller copy) */
  conditionCue?: string;
  /** Geography or outlet shorthand (EU consignment, US auction floor, Japan Yahoo, etc.) */
  locationOrChannel?: string;
  transactionTypeGuess?: ComparableTransactionTypeGuess;
  /** One concise line tying this comp to the wizard inputs or tier hint */
  relevanceExplanation?: string;
  matchTier?: ComparableMatchTier;
  /** Optional HTTPS thumbnail when the cited evidence page has a stable image URL */
  imageUrl?: string;
}